% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dws2perf.R
\name{dws2perf}
\alias{dws2perf}
\title{dws2perf}
\usage{
dws2perf(observedFilePath, dwsFolderPath, startDate, endDate, captionDwsVar,
  captionObsVar, TW = "day", perfMatrixFileName = NA)
}
\arguments{
\item{observedFilePath}{File path containing observed data.}

\item{dwsFolderPath}{Folder path containing .DWS output files.}

\item{startDate}{Start date with format YYYY-MM-DD (e.g. "2004-05-18").}

\item{endDate}{End date with format YYYY-MM-DD (e.g. "2004-05-18").}

\item{captionDwsVar}{Variable caption as it appears in the .DWS output file.}

\item{captionObsVar}{variable caption as it appears in the observed file.}

\item{TW}{Time averaging window could be: "day" , or "week", or "month" or "year" or any multiple of them, e.g. "3 month".}

\item{perfMatrixFileName}{An arbitrary file name for saving performance matrix. Will not save if not assigned a value.}
}
\value{
A data frame containing model performance values.
}
\description{
Calculates performance matrix using observed data and simulated .DWS files for a selected time averaging window.
}
\examples{
\dontrun{
# Creating a copy of tutorial folder:
    getExampleFolder()
# Creating a list for setting inputs:
    globalInput <- inputGen()
# Setting parameter bounds:
    globalInput$apexPARM$Root_growth_soil[1] = 0.15
    globalInput$apexPARM$Root_growth_soil[2] = 0.2
    globalInput$apexPARM$Soil_evap_plant_cover[1] = 0
    globalInput$apexPARM$Soil_evap_plant_cover[2] = 0.5
# Performing Monte Carlo simulation:
    input4SA <- APEXSENSUN::mc4APEX(globalInput)
# Calculating performance matrix:
    PerfMat <- dws2perf(observedFilePath ="Example/Observed_Files/observed.txt",
                    dwsFolderPath = "Example/Calculated_Outputs/DWS",
                    startDate="2002-01-01",
                    endDate="2003-01-01",
                    captionDwsVar="ET",captionObsVar="ET",TW="week")
}

}
