% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genParmRange.R
\name{genParmRange}
\alias{genParmRange}
\title{genParmRange}
\usage{
genParmRange(parmRangeFile)
}
\arguments{
\item{parmRangeFile}{File name associated to parameter range file containing parameters inside the PATM.dat file.}
}
\value{
Void
}
\description{
Creates an empty dataframe for setting lower and upper limits of uncertain model parameters inside the PARM.dat file
}
\examples{
\dontrun{
  genParmRange("parmRangeFile.txt") #Note: Used only if the inputs are to be read through text file.
  }
}
