% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExampleFolder.R
\name{getExampleFolder}
\alias{getExampleFolder}
\title{getExampleFolder}
\usage{
getExampleFolder(folderName = "Example")
}
\arguments{
\item{folderName}{A string representing tutorial folder name.}
}
\value{
Void/Creates a copy of the tutorial project inside working directory.
}
\description{
Creates a copy of the tutorial project inside working directory.
}
\examples{
\dontrun{
getExampleFolder()
}
}
