% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx2parm.R
\name{idx2parm}
\alias{idx2parm}
\title{idx2parm}
\usage{
idx2parm(xMat, idxAccept)
}
\arguments{
\item{xMat}{A data.frame object containing prior parameter distributions(i.e. prior distribution can be retrived from the output of MC4APEX function).}

\item{idxAccept}{An integer vector containing accepted simulation numbers.}
}
\value{
A data.frame object containing posterior distribution.
}
\description{
Derives posterior distribution using prior distribution and vector of accepted indices.
}
\examples{
\dontrun{
# Creating a copy of tutorial folder:
    getExampleFolder()
# Creating a list for setting inputs:
    globalInput <- inputGen()
# Setting parameter bounds:
    globalInput$apexPARM$Root_growth_soil[1] = 0.15
    globalInput$apexPARM$Root_growth_soil[2] = 0.2
    globalInput$apexPARM$Soil_evap_plant_cover[1] = 0
    globalInput$apexPARM$Soil_evap_plant_cover[2] = 0.5
# Performing Monte Carlo simulation:
    input4SA <- APEXSENSUN::mc4APEX(globalInput)
# Creating posterior distribution:
    posteriorParms <- idx2parm(xMat =input4SA$gsaObject$X,
                      idxAccept = c(20,23,56,776))
        }
}
