% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx2statTimeseries.R
\name{idx2statTimeseries}
\alias{idx2statTimeseries}
\title{idx2statTimeseries}
\usage{
idx2statTimeseries(dwsFolderPath, startDate, endDate, captionDwsVar,
  TW = "day", acceptedIndices = "all", probs = c(0.025, 0.5, 0.975))
}
\arguments{
\item{dwsFolderPath}{Folder path for .DWS output folder.}

\item{startDate}{Start date with format YYYY-MM-DD (e.g. "2002-01-01").}

\item{endDate}{End date with format YYYY-MM-DD (e.g. "2003-01-01").}

\item{captionDwsVar}{Variable caption as it appears in the .DWS output file.}

\item{TW}{Time averaging window could be: "day" , or "week", or "month" or "year" or any multiple of them, e.g. "3 month".}

\item{acceptedIndices}{"all" for all files or An integer vector containing accepted simulation numbers.}

\item{probs}{Vector of probabilities for which quantiles are calculated at each time step, e.g. c(0.025,0.5,0.975).}
}
\value{
A data frame containing mean and quantile time series for selected probabilities.
}
\description{
Generates time series statistics for accepted simulations using selected probabilities.
}
\examples{
\dontrun{
# Creating a copy of tutorial folder:
    getExampleFolder()
# Creating a list for setting inputs:
    globalInput <- inputGen()
# Setting parameter bounds:
    globalInput$apexPARM$Root_growth_soil[1] = 0.15
    globalInput$apexPARM$Root_growth_soil[2] = 0.2
    globalInput$apexPARM$Soil_evap_plant_cover[1] = 0
    globalInput$apexPARM$Soil_evap_plant_cover[2] = 0.5
# Performing Monte Carlo simulation:
    input4SA <- APEXSENSUN::mc4APEX(globalInput)
# Calculating stat time series:
    statTimeSeries <- idx2statTimeseries(dwsFolderPath = "Example/Calculated_Outputs/DWS/",
                                         startDate = "2002-01-01",
                                         endDate="2003-01-01",
                                         captionDwsVar="ET", TW = "week",
                                         acceptedIndices = "all",
                                         probs = c(0.025,0.5,0.85,0.975))
}

}
