% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idx2Timeseries.R
\name{idx2Timeseries}
\alias{idx2Timeseries}
\title{idx2Timeseries}
\usage{
idx2Timeseries(dwsFolderPath, startDate, endDate, captionDwsVar, TW = "day",
  acceptedIndices = "all")
}
\arguments{
\item{dwsFolderPath}{Folder path for .DWS output folder.}

\item{startDate}{Start date with format YYYY-MM-DD (e.g. "2002-01-01").}

\item{endDate}{End date with format YYYY-MM-DD (e.g. "2003-01-01").}

\item{captionDwsVar}{Variable caption as it appears in the .DWS output file.}

\item{TW}{Time averaging window could be: "day" , or "week", or "month" or "year" or any multiple of them, e.g. "3 month".}

\item{acceptedIndices}{An integer vector containing accepted simulation numbers.}
}
\value{
A dataframe containing aggregated time series for the selected TW.
}
\description{
Extracts output time series corresponding to accepted simulation. indices with a selected time window.
}
\examples{
\dontrun{
# Creating a copy of tutorial folder:
    getExampleFolder()
# Creating a list for setting inputs:
    globalInput <- inputGen()
# Setting parameter bounds:
    globalInput$apexPARM$Root_growth_soil[1] = 0.15
    globalInput$apexPARM$Root_growth_soil[2] = 0.2
    globalInput$apexPARM$Soil_evap_plant_cover[1] = 0
    globalInput$apexPARM$Soil_evap_plant_cover[2] = 0.5
# Performing Monte Carlo simulation:
    input4SA <- APEXSENSUN::mc4APEX(globalInput)
# Extracting time series
  extractedTimeseries <- idx2Timeseries(dwsFolderPath = "Example/Calculated_Outputs/DWS",
                                        startDate="2002-01-01", endDate="2003-01-01",
                                        captionDwsVar = "ET",TW="month",
                                        acceptedIndices=c(1,15,10,8))
}

}
