% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputGen.R
\name{inputGen}
\alias{inputGen}
\title{inputGen}
\usage{
inputGen(inputs)
}
\arguments{
\item{inputs}{Not required for setting simulation and SA parameters through a List object. Othereise, a File name for setting parameters through text file.}
}
\value{
A template list object for setting simulation and SA parameters, or Void if the inputs are to be read through text file (i.e. if 'input' is set to a value).
}
\description{
Generates a List object or text file for setting simulation and SA parameters.
}
\examples{
\dontrun{
input_list <- inputGen()   # Creates a List object For setting parameters.
inputGen("Input_File.txt") # Creates a text file For setting parameters (Deprecated) .
}
}
