% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc4APEX.R
\name{mc4APEX}
\alias{mc4APEX}
\title{mc4APEX}
\usage{
mc4APEX(input, usingTextFile = FALSE, savingInput4SA = TRUE)
}
\arguments{
\item{input}{A list or a file name containing simulation and SA parameters.}

\item{usingTextFile}{A logical variable specifying weather to read input from a list object or a text file.}

\item{savingInput4SA}{A logical variable specifying weather to save outputs.}
}
\value{
A list object containing uncertainty and sensitivity analysis settings and prior parameter distributions.
}
\description{
Performs a Monte Carlo simulation using the settting in an input list or input text file generated from a call to "inputGen" function.
}
\examples{
\dontrun{
# Creating a copy of tutorial folder:
    getExampleFolder()
# Creating a list for setting inputs:
    globalInput <- inputGen()
# Setting parameter bounds:
    globalInput$apexPARM$Root_growth_soil[1] = 0.15
    globalInput$apexPARM$Root_growth_soil[2] = 0.2
    globalInput$apexPARM$Soil_evap_plant_cover[1] = 0
    globalInput$apexPARM$Soil_evap_plant_cover[2] = 0.5
# Performing Monte Carlo simulation:
    input4SA <- APEXSENSUN::mc4APEX(globalInput)
}
}
