% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputIdxCombiner.R
\name{outputIdxCombiner}
\alias{outputIdxCombiner}
\title{outputIdxCombiner}
\usage{
outputIdxCombiner(...)
}
\arguments{
\item{...}{Any number of integer vectors.}
}
\value{
An interger vector representing idices that are common among all vectors.
}
\description{
Pulls out common indices from a number of integer vectors.
}
\examples{
\dontrun{
#Outputs an integer vector containing 7,8,9,10.
  combined_indices <- outputIdxCombiner(c(1:10),c(7:13))
  }
}
