% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf2idx.R
\name{perf2idx}
\alias{perf2idx}
\title{perf2idx}
\usage{
perf2idx(perfMatrix, lowLimit, upLimit)
}
\arguments{
\item{perfMatrix}{A data.frame object containing performance matrix (i.e. output of dws2perf function).}

\item{lowLimit}{A vector contatining the lower limits for RMSE, NASH and PBIAS, Mean e.g. c(0,0,-25,0).}

\item{upLimit}{A vector contatining the upper limits for RMSE, NASH and PBIAS, Mean, e.g. c(10,1,+25,5).}
}
\value{
A data frame containing accepted simulations and their performance measure values.
}
\description{
Renders simulation numbers meeting selected performance criteria.
}
\examples{
\dontrun{
# Creating a copy of tutorial folder:
    getExampleFolder()
# Creating a list for setting inputs:
    globalInput <- inputGen()
# Setting parameter bounds:
    globalInput$apexPARM$Root_growth_soil[1] = 0.15
    globalInput$apexPARM$Root_growth_soil[2] = 0.2
    globalInput$apexPARM$Soil_evap_plant_cover[1] = 0
    globalInput$apexPARM$Soil_evap_plant_cover[2] = 0.5
# Performing Monte Carlo simulation:
    input4SA <- APEXSENSUN::mc4APEX(globalInput)
# Calculating performance matrix:
    PerfMat <- dws2perf(observedFilePath ="Example/Observed_Files/observed.txt",
                    dwsFolderPath = "Example/Calculated_Outputs/DWS",
                    startDate="2002-01-01",
                    endDate="2003-01-01",
                    captionDwsVar="ET",captionObsVar="ET",TW="week")
# Extracting simulations meeting performance criteria:
    acceptedSimulations <- perf2idx(perfMatrix = PerfMat,
                                    lowLimit = c(0, 0, -2, 0),
                                    upLimit = c(10, 1, 2, 5))
}
}
