% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa4APEX.R
\name{sa4APEX}
\alias{sa4APEX}
\title{sa4APEX}
\usage{
sa4APEX(input, input4SA, usingTextFile = FALSE)
}
\arguments{
\item{input}{A list object (or a file name) containing simulation and SA parameters.}

\item{input4SA}{A list containing inputs for SA (i.e. output of a call to mc4APEX function)}

\item{usingTextFile}{A logical variable (i.e. either TRUE or FALSE) specifying weather to read "Input" from a List object or a text file.}
}
\value{
Void (i.e., writes sensitivity results in text files).
}
\description{
Performs sensitivity analysis for RMSE, NASH, PBIAS, and MEAN perofrmance measures.
}
\examples{
\dontrun{
# Creating a copy of tutorial folder:
    getExampleFolder()
# Creating a list for setting inputs:
    globalInput <- inputGen()
# Setting parameter bounds:
    globalInput$apexPARM$Root_growth_soil[1] = 0.15
    globalInput$apexPARM$Root_growth_soil[2] = 0.2
    globalInput$apexPARM$Soil_evap_plant_cover[1] = 0
    globalInput$apexPARM$Soil_evap_plant_cover[2] = 0.5
# Performing Monte Carlo simulation:
    input4SA <- APEXSENSUN::mc4APEX(globalInput)
# Performing sensitivity analysis:
    sa4APEX(globalInput,input4SA = input4SA)
    }
}
